+++ src/mem.c
@@ -3,8 +3,8 @@
 //
 #include "mem.h"
 
-uint64_t registers[32];
-uint64_t program_counter = 0;
+uint32_t registers[32];
+uint32_t program_counter = 0;
 
 static uint8_t rom_ptr[ROM_SIZE];
 static uint8_t ram_ptr[RAM_SIZE];
@@ -13,7 +13,7 @@ uint8_t *mem_get_rom_ptr() {
     return rom_ptr;
 }
 
-uint8_t mem_read_b(uint64_t addr) {
+uint8_t mem_read_b(uint32_t addr) {
     if (addr >= ROM_START_ADDR && addr < ROM_START_ADDR + ROM_SIZE) {
         return rom_ptr[addr - ROM_START_ADDR];
     } else if (addr >= RAM_START_ADDR && addr < RAM_START_ADDR + RAM_SIZE) {
@@ -24,7 +24,7 @@ uint8_t mem_read_b(uint64_t addr) {
     }
 }
 
-uint16_t mem_read_h(uint64_t addr) {
+uint16_t mem_read_h(uint32_t addr) {
     if (addr >= ROM_START_ADDR && addr + 2 < ROM_START_ADDR + ROM_SIZE) {
         return rom_ptr[addr - ROM_START_ADDR] | (rom_ptr[addr - ROM_START_ADDR + 1] << 8);
     } else if (addr >= RAM_START_ADDR && +2 < RAM_START_ADDR + RAM_SIZE) {
@@ -35,7 +35,7 @@ uint16_t mem_read_h(uint64_t addr) {
     }
 }
 
-uint32_t mem_read_w(uint64_t addr) {
+uint32_t mem_read_w(uint32_t addr) {
     if (addr >= ROM_START_ADDR && addr + 4 < ROM_START_ADDR + ROM_SIZE) {
         return
                 rom_ptr[addr - ROM_START_ADDR] |
